requirements.txt
# Core AI and ML dependencies
anthropic>=0.25.0
chromadb>=0.4.22
sentence-transformers>=2.5.1
docling>=1.8.0
pypdf>=3.17.4
langchain-text-splitters>=0.0.1

# Web frameworks
flask>=3.0.0
flask-cors>=4.0.0
streamlit>=1.31.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.8.0
numpy>=1.26.4
pandas>=2.2.2

# Development and testing
pytest>=8.0.0
black>=24.1.1
flake8>=7.0.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# openai==1.3.0  # For OpenAI integration
# tiktoken==0.5.1  # For token counting
# python-multipart==0.0.6  # For file uploads
# jinja2==3.1.2  # For templating
# markdown==3.5.1  # For markdown processing
# requests==2.31.0  # For HTTP requests
